+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Bonus feats from changes weren't showing in the sources list
 - Condition icons wouldn't show up on linked tokens (#867)
 - Fine creatures were too small, and thus made changes to token presets below Medium size (may need manual alterations to previous tokens, if desired) (#864)
+- The setting for when low light vision should be applied was not working (#849)
 
 ### Changelog